Shader implementation (WIP)
Re-implemented the shader features from Daemian. Unsure if they all work
properly. Seeing as the modloader finds the default game, it attempts to
replace the default ones with the default ones. Should be able to
disable that through the modloader.cfg file soon. Anyway, here's an
early version with some extra features.